Added Documentation.txt
I have altered the code somewhat, to begin initial testing of my possibly soon-to-be deleted function, dateParser. The main thing I have added is Documentation.txt. Documentation.txt will be where other developers can find lengthier comments as well as information on the different sections of the code that I am working on or adding.